requirements.txt
pytesseract==0.3.6
progress==1.5
Pillow==6.2.2